+++ project-base/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - LegalConditionsSetting: added privacy policy article selection (@stanoMilan)
     - customers need to agree with privacy policy while registring, sending contact form and completing order process
 - SubscriptionFormType: added required privacy policy agreement checkbox (@simara-svatopluk)
+- subscription form: added link to privacy policy agreement article (@simara-svatopluk)
 
 ### Changed
 - cache is cleared before PHPUnit tests only when run via [Phing targets](docs/introduction/phing-targets.md), not when run using `phpunit` directly (@PetrHeinz)